demo_project/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10
    # via django
certifi==2020.6.20
    # via requests
chardet==3.0.4
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django-crispy-forms==1.11.0
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django==3.1.1
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   drf-yasg
djangorestframework-csv==2.1.0
    # via -r requirements.in
djangorestframework-yaml==2.0.0
    # via -r requirements.in
djangorestframework==3.11.1
    # via
    #   -r requirements.in
    #   djangorestframework-csv
    #   drf-renderer-xlsx
    #   drf-yasg
drf-renderer-xlsx==0.3.9
    # via -r requirements.in
drf-writable-nested==0.6.1
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
et-xmlfile==1.0.1
    # via openpyxl
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jdcal==1.4.1
    # via openpyxl
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
openpyxl==3.0.6
    # via drf-renderer-xlsx
packaging==20.9
    # via drf-yasg
pyparsing==2.4.7
    # via packaging
pytz==2020.1
    # via django
pyyaml==5.4.1
    # via djangorestframework-yaml
requests==2.24.0
    # via
    #   -r requirements.in
    #   coreapi
ruamel.yaml==0.16.12
    # via drf-yasg
six==1.15.0
    # via djangorestframework-csv
sqlparse==0.3.1
    # via django
unicodecsv==0.14.1
    # via djangorestframework-csv
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.25.10
    # via requests
